* if resolver dies, whine about this in log and disable ip resolution
PR: 8

Is MFC worth doing?


git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@3400 1738b366-ca78-dd11-ac90-000e8e03bdb1

+++ src/sog/resolver.c
@@ -23,7 +23,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $Id: resolver.c,v 1.19 2001-08-28 16:37:42 avn Exp $
+ * $Id: resolver.c,v 1.20 2001-11-21 18:30:54 avn Exp $
  */
 
 #include <sys/types.h>
@@ -122,8 +122,10 @@ resolver_destroy(void)
 {
 	fclose(rfin);
 	fclose(rfout);
-	kill(rpid, SIGTERM);
-	wait(NULL);
+	if (rpid > 0) {
+		kill(rpid, SIGTERM);
+		wait(NULL);
+	}
 }
 
 void
